+++ src/app/views/ChannelsDetail.svelte
@@ -5,9 +5,7 @@
   import {DIRECT_MESSAGE} from "@welshman/util"
   import {formatTimestamp} from "src/util/misc"
   import Channel from "src/partials/Channel.svelte"
-  import Content from "src/partials/Content.svelte"
   import Popover from "src/partials/Popover.svelte"
-  import Modal from "src/partials/Modal.svelte"
   import Anchor from "src/partials/Anchor.svelte"
   import PersonCircles from "src/app/shared/PersonCircles.svelte"
   import PersonAbout from "src/app/shared/PersonAbout.svelte"
@@ -23,7 +21,6 @@
     markChannelRead,
     getChannelIdFromEvent,
     listenForMessages,
-    sortEventsDesc,
     ensureMessagePlaintext,
   } from "src/engine"
 
@@ -46,34 +43,8 @@
       return sendLegacyMessage(channelId, content)
     }
 
-    const [message] = sortEventsDesc($messages || [])
-
-    if (!message || message?.kind === 4) {
-      confirmMessage = content
-    } else {
-      sendMessage(channelId, content)
-    }
+    sendMessage(channelId, content)
   }
-
-  const confirmNip04 = () => {
-    sendLegacyMessage(channelId, confirmMessage)
-    confirmMessage = null
-  }
-
-  const confirmNip44 = () => {
-    sendMessage(channelId, confirmMessage)
-    confirmMessage = null
-  }
-
-  const abortMessage = () => {
-    if (confirmMessage) {
-      ctrl.setMessage(confirmMessage)
-    }
-
-    confirmMessage = null
-  }
-
-  let confirmMessage, ctrl
 
   onMount(() => {
     markChannelRead(channelId)
@@ -88,7 +59,7 @@
   document.title = `Direct Messages`
 </script>
 
-<Channel {channelId} {pubkeys} bind:this={ctrl} messages={$messages} sendMessage={send} {initialMessage}>
+<Channel {pubkeys} messages={$messages} sendMessage={send} {initialMessage}>
   <div slot="header" class="flex h-16 items-start gap-4 overflow-hidden p-1 px-4">
     <div class="flex items-center gap-4 pt-1">
       <Anchor class="fa fa-arrow-left cursor-pointer text-2xl" href="/channels" />
@@ -134,20 +105,22 @@
       class:text-neutral-100={message.pubkey !== $session.pubkey}>
       {formatTimestamp(message.created_at)}
       {#if message.kind === 4}
-        <Popover>
-          <i slot="trigger" class="fa fa-unlock cursor-pointer text-neutral-200" />
+        <Popover triggerType="mouseenter">
+          <i slot="trigger" class="fa fa-unlock cursor-pointer text-neutral-400" />
           <p slot="tooltip">
             This message was sent using nostr's legacy DMs, which have a number of shortcomings.
             Read more <Anchor underline modal href="/help/nip-44-dms">here</Anchor>.
           </p>
         </Popover>
       {:else}
-        <Popover>
-          <i slot="trigger" class="fa fa-lock cursor-pointer text-neutral-200" />
+        <Popover triggerType="mouseenter">
+          <i slot="trigger" class="fa fa-lock cursor-pointer text-neutral-400" />
           <div slot="tooltip" class="flex flex-col gap-2">
             <p>
               This message was sent using nostr's new group chat specification, which solves several
-              problems with legacy DMs. Read more <Anchor underline modal href="/help/nip-44-dms">here</Anchor>.
+              problems with legacy DMs. Read more <Anchor underline modal href="/help/nip-44-dms"
+                >here</Anchor
+              >.
             </p>
             <p>
               Note that these messages are not yet universally supported. Make sure the person
@@ -159,29 +132,3 @@
     </small>
   </div>
 </Channel>
-
-{#if confirmMessage}
-  <Modal onEscape={abortMessage}>
-    <Content size="lg">
-      <p class="flex items-center gap-4 text-xl">
-        <i class="fa fa-info-circle" /> Auto-upgrade notice
-      </p>
-      <p>
-        This conversation has not yet been upgraded to use <Anchor
-          underline
-          modal
-          href="/help/nip-44-dms">new-style DMs</Anchor
-        >.
-      </p>
-      <p>
-        You should make sure @{displayProfileByPubkey(pubkeys[0])} is using a compatible nostr client,
-        or you can choose to send an old-style message instead.
-      </p>
-      <p>How would you like to send this message?</p>
-      <div class="flex flex-col gap-2 py-4 sm:flex-row">
-        <Anchor button on:click={confirmNip04}>Send using Legacy DMs</Anchor>
-        <Anchor button accent on:click={confirmNip44}>Send using NIP 44</Anchor>
-      </div>
-    </Content>
-  </Modal>
-{/if}